+++ Glib/data_loader/Alchemy_dataset.py
@@ -2,6 +2,7 @@ import os.path as osp
 import torch
 from torch_geometric.data import Data
 from torch_geometric.data import InMemoryDataset
+from torchvision.datasets import DatasetFolder
 from rdkit import Chem
 from rdkit.Chem import ChemicalFeatures
 from rdkit import RDConfig
@@ -21,11 +22,21 @@ class TencentAlchemyDataset(InMemoryDataset):
     fdef_name = osp.join(RDConfig.RDDataDir, 'BaseFeatures.fdef')
     chem_feature_factory = ChemicalFeatures.BuildFeatureFactory(fdef_name)
 
-    def __init__(self, root, mode='dev', transform=None, pre_transform=None, pre_filter=None):
+    def __init__(
+        self, root, mode='dev',
+        extenstion=None,
+        transform=None,
+        pre_transform=None,
+        pre_filter=None):
+
         self.mode = mode
         assert mode in _urls
         super(TencentAlchemyDataset, self).__init__(root, transform, pre_transform, pre_filter)
         self.data, self.slices = torch.load(self.processed_paths[0])
+        if self.mode == 'dev':
+            self.stratify = torch.load(self.processed_paths[1])
+        else:
+            self.stratify = None
 
     @property
     def raw_file_names(self):
@@ -33,7 +44,7 @@ class TencentAlchemyDataset(InMemoryDataset):
 
     @property
     def processed_file_names(self):
-        return 'TencentAlchemy_%s.pt' % self.mode
+        return ['TencentAlchemy_%s.pt' % self.mode, 'num_heavy_atoms_%s.pt' % self.mode] if self.mode == 'dev' else ['TencentAlchemy_%s.pt' % self.mode, ]
 
     def download(self):
         raise NotImplementedError('please download and unzip dataset from %s, and put it at %s' % (_urls[self.mode], self.raw_dir))
@@ -79,7 +90,7 @@ class TencentAlchemyDataset(InMemoryDataset):
 
     # sdf file reader for Alchemy dataset
     def sdf_graph_reader(self, sdf_file):
-
+        sdf_file = pathlib.Path(sdf_file)
         with open(sdf_file, 'r') as f:
             sdf_string = f.read()
         mol = Chem.MolFromMolBlock(sdf_string, removeHs=False)
@@ -136,24 +147,18 @@ class TencentAlchemyDataset(InMemoryDataset):
             self.target = pd.read_csv(self.raw_paths[1], index_col=0,
                     usecols=['gdb_idx',] + ['property_%d' % x for x in range(12)])
             self.target = self.target[['property_%d' % x for x in range(12)]]
-        sdf_dir = pathlib.Path(self.raw_paths[0])
+        
+        _dataset = DatasetFolder(root=self.raw_paths[0], loader=self.sdf_graph_reader,
+                                extensions='sdf', transform=self.pre_transform)
+
         data_list = []
-        class_vector = []
-        for atom_dir in sdf_dir.glob("atom_*"):
-            num_heavy_atoms = atom_dir.name.split('_')[1]
-            for sdf_file in atom_dir.glob("**/*.sdf"):
-                alchemy_data = self.sdf_graph_reader(sdf_file)
-                if alchemy_data is not None:
-                    data_list.append(alchemy_data)
-                    class_vector.append(num_heavy_atoms)
+        for alchemy_data, _ in _dataset:
+            data_list.append(alchemy_data)
 
         if self.pre_filter is not None:
             data_list = [data for data in data_list if self.pre_filter(data)]
 
-        if self.pre_transform is not None:
-            data_list = [self.pre_transform(data) for data in data_list]
-
         data, slices = self.collate(data_list)
         torch.save((data, slices), self.processed_paths[0])
-        torch.save(class_vector, self.processed_dir+'/'+'class_vector.pt')
-
+        if self.mode == 'dev':
+            torch.save(_dataset.targets, self.processed_paths[1])