+++ Glib/data_loader/data_loaders.py
@@ -13,10 +13,5 @@ class AlchemyDataLoader(BaseDataLoader):
         
         transform = None
         self.dataset = TencentAlchemyDataset(data_dir, mode, transform=transform)
-        
-        if training:
-            class_vector = torch.load(self.dataset.processed_dir+'/'+'class_vector.pt')
-        else:
-            class_vector = None
-        
-        super().__init__(self.dataset, batch_size, shuffle, validation_split, num_workers, stratify=class_vector)
+
+        super().__init__(self.dataset, batch_size, shuffle, validation_split, num_workers)